# api/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Buzz BD Agent API \u2014 SolCex Exchange",
    "description": "Autonomous BD intelligence and security scanning for crypto tokens. Token scoring, pipeline analytics, mining intel, Monte Carlo simulations, and BuzzShield security audits. All premium endpoints accept x402 micropayments via USDC on Base.",
    "version": "9.5.0",
    "contact": {
      "name": "Buzz BD Agent",
      "url": "https://buzzbd.ai",
      "email": "buzzbysolcex@gmail.com"
    },
    "x-guidance": "Use /premium/score for individual token risk scoring (pass token contract address + chain). Use /premium/pipeline for full pipeline analytics across all tracked tokens. Use /shield/scan for quick security checks and /shield/audit/full for comprehensive smart contract audits. All endpoints return JSON. Payment is handled automatically via x402 \u2014 your agent's wallet pays per-request in USDC on Base network."
  },
  "servers": [
    {
      "url": "https://api.buzzbd.ai",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/premium/sim": {
      "post": {
        "operationId": "runSimulation",
        "summary": "Monte Carlo simulation \u2014 multi-agent consensus on token trajectory",
        "description": "Run a MiroFish-style Monte Carlo simulation with 20 synthetic agents evaluating a token's price trajectory, risk factors, and consensus outlook.",
        "x-payment-info": {
          "price": "0.01",
          "currency": "USDC",
          "network": "base",
          "description": "Monte Carlo simulation run"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["contractAddress", "chain"],
                "properties": {
                  "contractAddress": {
                    "type": "string"
                  },
                  "chain": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Simulation results with agent consensus"
          },
          "402": {
            "description": "Payment required"
          }
        },
        "x-payment-protocols": ["exact"]
      }
    },
...
  }
}